chore: move share sheet component to folder
